requirements.txt
streamlit
pandas
PyMuPDF
sentence-transformers
faiss-cpu
matplotlib
requests